# TreckSec_API_Documentation.yaml
openapi: 3.0.0
info:
  version: '2.0'
  title: "TrekSec OpenAPI 3.0"
  description: API for a secure trekking.
  license:
    name: MIT
servers:
  - url: http://localhost:8080/api/v1
    description: Localhost
paths:
...
  /users/reset-password:
    post:
      description: >-
        gets the mail (used to send an email) to reset the password.
      summary: Send an email
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
                  description: 'Email address of the student'
      responses:
        '201':
          description: 'Email sent to the user that required the reset assword'
          content:
            application/json:
              schema:
                type: string
                description:  Mail sent message.
        '404':
          description: 'Email not found'
          content:
            application/json::
              schema:
                type: string
                description:  Email not found message.
              
    put:
      description: >-
        update the password of the user
      summary: Update the password
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - token
                - newPassword
                - newcheckPassword
              properties:
                token:
                  type: string
                  description: 'Token of the user'
                newPassword:
                  type: string
                  description: 'newPassword of the user'
                newcheckPassword:
                  type: string
                  description: 'Newcheck of the users password'
      responses:
        '201':
          description: 'Password changed successfully'
          content:
            application/json:
              schema:
                type: string
                description:  Mail sent message.
        '400':
          description: 'Error message sent, with the actual error'
          content:
            application/json:
              schema:
                type: string
                description:  error message.
        '500':
          description: 'Si è verificato un errore durante il reset della password. Riprova più tardi'
          content:
            application/json:
              schema:
                type: string
                description:  Si è verificato un errore durante il reset della password. Riprova più tardi message.
...